+++ src/command_modules/azure-cli-eventhubs/azure/cli/command_modules/eventhubs/_utils.py
@@ -5,7 +5,7 @@
 
 
 def accessrights_converter(accessrights):
-    from azure.mgmt.eventhub.models.event_hub_management_client_enums import AccessRights
+    from azure.mgmt.eventhub.models import AccessRights
     accessrights_new = []
     if 'Send' in accessrights:
         accessrights_new.append(AccessRights.send)