+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_acl_scenarios.py
@@ -6,7 +6,7 @@
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, StorageAccountPreparer, JMESPathCheck, NoneCheck,
                                api_version_constraint)
 from azure.cli.core.profiles import ResourceType
-from .storage_test_util import StorageScenarioMixin
+from ..storage_test_util import StorageScenarioMixin
 
 
 @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2016-12-01')
@@ -52,7 +52,7 @@ class StorageAccessControlListTests(StorageScenarioMixin, ScenarioTest):
                          container_id_parameter).assert_with_checks(JMESPathCheck('start', '2016-01-01T00:00:00+00:00'))
         self.storage_cmd('storage {} policy show {} -n test3', account_info, service_type,
                          container_id_parameter).assert_with_checks(
-            JMESPathCheck('expiry', '2018-01-01T00:00:00+00:00'))
+                             JMESPathCheck('expiry', '2018-01-01T00:00:00+00:00'))
         self.storage_cmd('storage {} policy show {} -n test4', account_info, service_type,
                          container_id_parameter).assert_with_checks(JMESPathCheck('start', '2016-01-01T00:00:00+00:00'),
                                                                     JMESPathCheck('expiry',