+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,9 +2,15 @@
 
 Release History
 ===============
+0.2.4
++++++
+* appservice plan, webapp & function app updated to use latest python websites SDK version
+
 0.2.3
 +++++
+* support CORS on functionapp & webapp
 * arm tag support on create commands
+* `webapp/functionapp identity show`: exception handling to exit with code 3 upon a missing resource for consistency
 
 0.2.2
 +++++