+++ src/command_modules/azure-cli-cognitiveservices/HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ===============
 
+0.2.2
++++++
+* Add new command `az cognitiveservices account list-usage`
+* Add new command `az cognitiveservices account list-kinds`
+* Add new command `az cognitiveservices account list`
+* Deprecate `az cognitiveservices list`
+
+0.2.1
++++++
+* Add new parameter --api-properties, which is required for creating some of the services.
+
 0.2.0
 +++++
 * BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.