+++ src/command_modules/azure-cli-iotcentral/HISTORY.rst
@@ -3,7 +3,14 @@
 Release History
 ===============
 
+0.1.2
++++++
+* Minor fixes
+
+0.1.1
++++++
+* Minor fixes
+
 0.1.0
 +++++
-
 * Initial release of IoT Central module.